refactor: browserwindow.tsx
